only remove homologs if the homologs file is present

+++ helix/patchman/run_patchman.py
@@ -91,23 +91,24 @@ def main():
 
     # Remove homologs
     homologs = []
-    with open(os.path.join(workspace.focus_dir, 'homologs'), 'r') as f:
-        for line in f:
-            homologs.append(line.strip().lower())
-    matches = glob.glob('*_matches')
-    fout = open('db_list_nohom', 'w')
-    removed = open('removed_psds', 'w')
-    with open('db_list', 'r') as f:
-        for line in f:
-            pdbid = line.split('/')[-1].split('_')[0]
-            if pdbid not in homologs:
-                fout.write(line)
-            else:
-                removed.write(line)
-    fout.close()
-    removed.close()
-    os.remove('db_list')
-    os.rename('db_list_nohom', 'db_list')
+    if os.path.exists(os.path.join(workspace.focus_dir, 'homologs')):
+        with open(os.path.join(workspace.focus_dir, 'homologs'), 'r') as f:
+            for line in f:
+                homologs.append(line.strip().lower())
+        matches = glob.glob('*_matches')
+        fout = open('db_list_nohom', 'w')
+        removed = open('removed_psds', 'w')
+        with open('db_list', 'r') as f:
+            for line in f:
+                pdbid = line.split('/')[-1].split('_')[0]
+                if pdbid not in homologs:
+                    fout.write(line)
+                else:
+                    removed.write(line)
+        fout.close()
+        removed.close()
+        os.remove('db_list')
+        os.rename('db_list_nohom', 'db_list')
 
     # Run MASTER for all motifs
     pds_list = glob.glob('*pds')
